requirements.txt
python-telegram-bot[webhooks]>=20.5
openai>=1.0.0
python-dotenv
